fix: Run formatter

+++ src/Theopetra/StakingDistributor.sol
@@ -11,68 +11,64 @@ import "../Interfaces/IERC20.sol";
 import "../Interfaces/IDistributor.sol";
 
 contract StakingDistributor is IDistributor, TheopetraAccessControlled {
-    using SafeMath for uint;
+    using SafeMath for uint256;
     using SafeERC20 for IERC20;
-
-
 
     /* ====== VARIABLES ====== */
 
     address public immutable THEO;
     address public immutable treasury;
 
-    uint public immutable epochLength;
-    uint public nextEpochBlock;
+    uint256 public immutable epochLength;
+    uint256 public nextEpochBlock;
 
-    mapping( uint => Adjust ) public adjustments;
-
+    mapping(uint256 => Adjust) public adjustments;
 
     /* ====== STRUCTS ====== */
 
     struct Info {
-        uint rate; // in ten-thousandths ( 5000 = 0.5% )
+        uint256 rate; // in ten-thousandths ( 5000 = 0.5% )
         address recipient;
     }
     Info[] public info;
 
     struct Adjust {
         bool add;
-        uint rate;
-        uint target;
+        uint256 rate;
+        uint256 target;
     }
-
-
 
     /* ====== CONSTRUCTOR ====== */
 
-    constructor( address _treasury, address _theo, uint _epochLength, uint _nextEpochBlock, ITheopetraAuthority _authority ) TheopetraAccessControlled(_authority) {
-        require( _treasury != address(0) );
+    constructor(
+        address _treasury,
+        address _theo,
+        uint256 _epochLength,
+        uint256 _nextEpochBlock,
+        ITheopetraAuthority _authority
+    ) TheopetraAccessControlled(_authority) {
+        require(_treasury != address(0));
         treasury = _treasury;
-        require( _theo != address(0) );
+        require(_theo != address(0));
         THEO = _theo;
         epochLength = _epochLength;
         nextEpochBlock = _nextEpochBlock;
     }
-
-
 
     /* ====== PUBLIC FUNCTIONS ====== */
 
     /**
         @notice send epoch reward to staking contract
      */
-    function distribute() external override returns ( bool ) {
-        if ( nextEpochBlock <= block.number ) {
-            nextEpochBlock = nextEpochBlock.add( epochLength ); // set next epoch block
+    function distribute() external override returns (bool) {
+        if (nextEpochBlock <= block.number) {
+            nextEpochBlock = nextEpochBlock.add(epochLength); // set next epoch block
 
             // distribute rewards to each recipient
-            for ( uint i = 0; i < info.length; i++ ) {
-                if ( info[ i ].rate > 0 ) {
-                    ITreasury( treasury ).mint( // mint and send from treasury
-                        info[ i ].recipient,
-                        nextRewardAt( info[ i ].rate )
-                    );
-                    adjust( i ); // check for adjustment
+            for (uint256 i = 0; i < info.length; i++) {
+                if (info[i].rate > 0) {
+                    ITreasury(treasury).mint(info[i].recipient, nextRewardAt(info[i].rate)); // mint and send from treasury
+                    adjust(i); // check for adjustment
                 }
             }
             return true;
@@ -81,31 +77,31 @@ contract StakingDistributor is IDistributor, TheopetraAccessControlled {
         }
     }
 
-
-
     /* ====== INTERNAL FUNCTIONS ====== */
 
     /**
         @notice increment reward rate for collector
      */
-    function adjust( uint _index ) internal {
-        Adjust memory adjustment = adjustments[ _index ];
-        if ( adjustment.rate != 0 ) {
-            if ( adjustment.add ) { // if rate should increase
-                info[ _index ].rate = info[ _index ].rate.add( adjustment.rate ); // raise rate
-                if ( info[ _index ].rate >= adjustment.target ) { // if target met
-                    adjustments[ _index ].rate = 0; // turn off adjustment
+    function adjust(uint256 _index) internal {
+        Adjust memory adjustment = adjustments[_index];
+        if (adjustment.rate != 0) {
+            if (adjustment.add) {
+                // if rate should increase
+                info[_index].rate = info[_index].rate.add(adjustment.rate); // raise rate
+                if (info[_index].rate >= adjustment.target) {
+                    // if target met
+                    adjustments[_index].rate = 0; // turn off adjustment
                 }
-            } else { // if rate should decrease
-                info[ _index ].rate = info[ _index ].rate.sub( adjustment.rate ); // lower rate
-                if ( info[ _index ].rate <= adjustment.target ) { // if target met
-                    adjustments[ _index ].rate = 0; // turn off adjustment
+            } else {
+                // if rate should decrease
+                info[_index].rate = info[_index].rate.sub(adjustment.rate); // lower rate
+                if (info[_index].rate <= adjustment.target) {
+                    // if target met
+                    adjustments[_index].rate = 0; // turn off adjustment
                 }
             }
         }
     }
-
-
 
     /* ====== VIEW FUNCTIONS ====== */
 
@@ -114,8 +110,8 @@ contract StakingDistributor is IDistributor, TheopetraAccessControlled {
         @param _rate uint
         @return uint
      */
-    function nextRewardAt( uint _rate ) public view override returns ( uint ) {
-        return IERC20( THEO ).totalSupply().mul( _rate ).div( 1000000 );
+    function nextRewardAt(uint256 _rate) public view override returns (uint256) {
+        return IERC20(THEO).totalSupply().mul(_rate).div(1000000);
     }
 
     /**
@@ -123,17 +119,15 @@ contract StakingDistributor is IDistributor, TheopetraAccessControlled {
         @param _recipient address
         @return uint
      */
-    function nextRewardFor( address _recipient ) public view override returns ( uint ) {
-        uint reward;
-        for ( uint i = 0; i < info.length; i++ ) {
-            if ( info[ i ].recipient == _recipient ) {
-                reward = nextRewardAt( info[ i ].rate );
+    function nextRewardFor(address _recipient) public view override returns (uint256) {
+        uint256 reward;
+        for (uint256 i = 0; i < info.length; i++) {
+            if (info[i].recipient == _recipient) {
+                reward = nextRewardAt(info[i].rate);
             }
         }
         return reward;
     }
-
-
 
     /* ====== POLICY FUNCTIONS ====== */
 
@@ -142,12 +136,9 @@ contract StakingDistributor is IDistributor, TheopetraAccessControlled {
         @param _recipient address
         @param _rewardRate uint
      */
-    function addRecipient( address _recipient, uint _rewardRate ) external override onlyPolicy() {
-        require( _recipient != address(0) );
-        info.push( Info({
-            recipient: _recipient,
-            rate: _rewardRate
-        }));
+    function addRecipient(address _recipient, uint256 _rewardRate) external override onlyPolicy {
+        require(_recipient != address(0));
+        info.push(Info({ recipient: _recipient, rate: _rewardRate }));
     }
 
     /**
@@ -155,10 +146,10 @@ contract StakingDistributor is IDistributor, TheopetraAccessControlled {
         @param _index uint
         @param _recipient address
      */
-    function removeRecipient( uint _index, address _recipient ) external override onlyPolicy() {
-        require( _recipient == info[ _index ].recipient );
-        info[ _index ].recipient = address(0);
-        info[ _index ].rate = 0;
+    function removeRecipient(uint256 _index, address _recipient) external override onlyPolicy {
+        require(_recipient == info[_index].recipient);
+        info[_index].recipient = address(0);
+        info[_index].rate = 0;
     }
 
     /**
@@ -168,11 +159,12 @@ contract StakingDistributor is IDistributor, TheopetraAccessControlled {
         @param _rate uint
         @param _target uint
      */
-    function setAdjustment( uint _index, bool _add, uint _rate, uint _target ) external override onlyPolicy() {
-        adjustments[ _index ] = Adjust({
-            add: _add,
-            rate: _rate,
-            target: _target
-        });
+    function setAdjustment(
+        uint256 _index,
+        bool _add,
+        uint256 _rate,
+        uint256 _target
+    ) external override onlyPolicy {
+        adjustments[_index] = Adjust({ add: _add, rate: _rate, target: _target });
     }
 }
